UI updates and bug progress updates
UI has been updated. EditText has been removed because it is not needed
and replaced with a TextView to show the user their unique id.  The bug
has been traced to the Android unique ID being passed to the encrypt
object was not what was expected and not a valid key. Next update will
include the fix.